Use docker info --format more consistently
At some point the output of `docker info` changed in a way that broke
the storage driver check (it starts with a space now).  This switches
to the `docker info --format` option in a few more places to get
output that is less likely to change again in the future.